# src/main/resources/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "restfun 风格接口\n这里面的所有接口访问都需要权限验证\n都要在请求头中加入”authorization: Bearer “\n在空格后面加token\n用户登陆成功后返回的头中会有”authorization: Bearer token“",
    "version": "0.0.9-SNAPSHOT",
    "title": "铭贝api开发规范",
    "contact": {
      "name": "作者",
      "url": "https://baike.baidu.com/search/none?word=%E9%83%AD%E9%9B%A8%E8%A1%A1&pn=0&rn=10&enc=utf8",
      "email": "2350871838@qq.com"
    }
  },
...
  "host": "39.108.114.59",
  "basePath": "/",
...
  "paths": {
...
    "/system/role": {
      "get": {
        "tags": [
          "系统角色"
        ],
        "summary": "获取所有的角色列表",
        "operationId": "getAllUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "令牌",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Role"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      },
      "post": {
        "tags": [
          "系统角色"
        ],
        "summary": "添加一个角色",
        "operationId": "addRoleUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "令牌",
            "required": true,
            "type": "string"
          },
          {
            "name": "name",
            "in": "query",
            "description": "权限名称",
            "required": true,
            "type": "string"
          },
          {
            "name": "nameZh",
            "in": "query",
            "description": "角色中文名称（角色描述）",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/返回对象"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
...
    "/system/role/{id}": {
      "put": {
        "tags": [
          "系统角色"
        ],
        "summary": "更新角色信息",
        "operationId": "updateRoleUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "令牌",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "description": "角色id",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "name",
            "in": "query",
            "description": "角色名",
            "required": false,
            "type": "string"
          },
          {
            "name": "nameZh",
            "in": "query",
            "description": "权限的中文名称（权限描述）",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/返回对象«string»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      },
      "delete": {
        "tags": [
          "系统角色"
        ],
        "summary": "删除一个角色",
        "operationId": "deleteRoleUsingDELETE",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "令牌",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/返回对象«string»"
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "deprecated": false
      }
    },
...
    "/user": {
      "get": {
        "tags": [
          "用户"
        ],
        "summary": "获取当前用户所属节点下的所有用户",
        "description": "如果自己也在该单位下那么也能查出自己，如果用户不能查看其他人账户将只返回他自己",
        "operationId": "getAllUsingGET_1",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "令牌",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/User"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      },
      "post": {
        "tags": [
          "用户"
        ],
        "summary": "注册用户",
        "description": "权限存在问题，关于权限字段不用传",
        "operationId": "logonUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "令牌",
            "required": true,
            "type": "string"
          },
          {
            "name": "depId",
            "in": "query",
            "description": "属于哪个单位（单位id）",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "name",
            "in": "query",
            "description": "真实姓名",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "query",
            "description": "用户密码",
            "required": true,
            "type": "string"
          },
          {
            "name": "phone",
            "in": "query",
            "description": "手机号",
            "required": false,
            "type": "string"
          },
          {
            "name": "sex",
            "in": "query",
            "description": "性别",
            "required": false,
            "type": "string"
          },
          {
            "name": "username",
            "in": "query",
            "description": "用户名",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/返回对象«string»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/user/{id}": {
      "get": {
        "tags": [
          "用户"
        ],
        "summary": "获取用户信息 根据用户id",
        "description": "并不是所有用户都能获取，有权限",
        "operationId": "getUserByIdUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "令牌",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "description": "用户Id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/返回对象«string»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      },
      "put": {
        "tags": [
          "用户"
        ],
        "summary": "更新用户资料(可以更新自己的资料也可以更新下属的资料)",
        "description": "也可以更改自己的资料\n权限存在问题，关于权限字段不用传",
        "operationId": "updateUserRoleUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "令牌",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "description": "用户id",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "depId",
            "in": "query",
            "description": "属于哪个单位（单位id）",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "name",
            "in": "query",
            "description": "真实姓名",
            "required": false,
            "type": "string"
          },
          {
            "name": "password",
            "in": "query",
            "description": "用户密码",
            "required": false,
            "type": "string"
          },
          {
            "name": "phone",
            "in": "query",
            "description": "手机号",
            "required": false,
            "type": "string"
          },
          {
            "name": "sex",
            "in": "query",
            "description": "性别",
            "required": false,
            "type": "string"
          },
          {
            "name": "username",
            "in": "query",
            "description": "用户名",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/返回对象«string»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      },
      "delete": {
        "tags": [
          "用户"
        ],
        "summary": "通过用户id删除用户",
        "operationId": "deleteUsingDELETE_1",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "令牌",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/返回对象«string»"
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "deprecated": false
      }
    }
  },
  "definitions": {
...
    "Role": {
      "type": "object",
      "required": [
        "name"
      ],
      "properties": {
        "name": {
          "type": "string",
          "description": "权限名称"
        },
        "nameZh": {
          "type": "string",
          "description": "权限的中文名称（权限描述）"
        }
      },
      "title": "Role"
    },
    "User": {
      "type": "object",
      "required": [
        "depId",
        "name",
        "password",
        "username"
      ],
      "properties": {
        "depId": {
          "type": "integer",
          "format": "int64",
          "description": "属于哪个单位（单位id）"
        },
        "name": {
          "type": "string",
          "description": "真实姓名"
        },
        "password": {
          "type": "string",
          "description": "用户密码"
        },
        "phone": {
          "type": "string",
          "description": "手机号"
        },
        "sex": {
          "type": "string",
          "description": "性别"
        },
        "username": {
          "type": "string",
          "description": "用户名"
        }
      },
      "title": "User",
      "description": "用户"
    },
...
    "返回对象": {
      "type": "object",
      "properties": {
        "message": {
          "type": "object",
          "description": "对应信息"
        },
        "status": {
          "type": "string",
          "description": "状态码",
          "enum": [
            "200:操作成功",
            "400:操作失败（已知错误）",
            "304:操作失败（未知错误）"
          ]
        }
      },
      "title": "返回对象",
      "description": "统一返回格式"
    },
...
    "返回对象«string»": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "description": "对应信息"
        },
        "status": {
          "type": "string",
          "description": "状态码",
          "enum": [
            "200:操作成功",
            "400:操作失败（已知错误）",
            "304:操作失败（未知错误）"
          ]
        }
      },
      "title": "返回对象«string»",
      "description": "统一返回格式"
    }
  }
}